third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGLineElement-svgdom-x2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'x2' property of the SVGLineElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS lineElement.x2.baseVal.value is 100
PASS lineElement.x2.baseVal.value is 200
PASS successfullyParsed is true

TEST COMPLETE
